% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd.R
\name{cmd}
\alias{cmd}
\alias{cmd_bash}
\alias{cmd_r}
\alias{cmd_docker}
\alias{cmd_asciinema}
\title{Command composition}
\usage{
cmd(command = NULL, args = NULL, ...)

cmd_bash(command = NULL)

cmd_r(command = NULL, vanilla = TRUE)

cmd_docker(command = NULL, image, remove = TRUE, ...)

cmd_asciinema(command = NULL, filename = NULL, overwrite = TRUE)
}
\arguments{
\item{command}{A string.}

\item{args}{A vector of strings.}

\item{...}{Strings.}

\item{vanilla}{A logical.}

\item{image}{A string.}

\item{remove}{A logical.}

\item{filename}{A string.}

\item{overwrite}{A logical.}
}
\value{
A vector of strings.
}
\description{
Command composition
}
