% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{cursor}
\alias{cursor}
\title{Get cursor position on screen}
\usage{
cursor(session)
}
\arguments{
\item{session}{A rexpect_session.}
}
\value{
A numerical vector with two values.
}
\description{
Starts at top-left corner (1, 1) and ends at bottom-right corner
(width, height)
}
