% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{ends_with_prompt}
\alias{ends_with_prompt}
\title{Does the contents of a pane end with prompt?}
\usage{
ends_with_prompt(session)
}
\arguments{
\item{session}{A rexpect_session.}
}
\description{
Does the contents of a pane end with prompt?
}
