% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect.R
\name{expect_prompt}
\alias{expect_prompt}
\title{Expect prompt}
\usage{
expect_prompt(session, interval = 0.5, timeout = NULL)
}
\arguments{
\item{session}{A rexpect_session.}

\item{interval}{Numerical. Time to wait between tries in seconds.
Default: 0.5.}

\item{timeout}{Numerical. Maximum amount of time to wait.
Default: \code{NULL}.}
}
\description{
Wait until the prompt appears on the last line.
}
