% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect.R
\name{expect_silence}
\alias{expect_silence}
\title{Expect silence}
\usage{
expect_silence(session, duration = 1, interval = 0.5, timeout = NULL)
}
\arguments{
\item{session}{A rexpect_session.}

\item{duration}{Numerical. Time output needs to remain unchanged in seconds.
Default: 1.}

\item{interval}{Numerical. Time to wait between tries in seconds.
Default: 0.5.}

\item{timeout}{Numerical. Maximum amount of time to wait.
Default: \code{NULL}.}
}
\description{
Wait until the output hasn't changed for a while.
}
