% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{get_prompt}
\alias{get_prompt}
\title{Get the prompt}
\usage{
get_prompt(session)
}
\arguments{
\item{session}{A rexpect_session.}
}
\value{
String containing a regular expression that matches all relevant
prompts.
}
\description{
Get the prompt
}
