% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{length.rexpect_session}
\alias{length.rexpect_session}
\title{Number of lines}
\usage{
\method{length}{rexpect_session}(x)
}
\arguments{
\item{x}{A rexpect_session.}
}
\value{
An integer.
}
\description{
Includes the current line.
}
