% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{prompt}
\alias{prompt}
\alias{prompt<-}
\alias{set_prompt}
\title{Prompt of an rexpect session}
\usage{
prompt(session)

prompt(session) <- value

set_prompt(session, prompt)
}
\arguments{
\item{session}{A rexpect_session.}

\item{prompt}{String containing a regular expression that matches all
relevant patterns.}
}
\value{
String containing a regular expression that matches all relevant
patterns.
}
\description{
Functions to get and set the prompt of an rexpect session. Note: this
function does not change the prompt of the command that is running.
}
