% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_screen}
\alias{read_screen}
\title{Read screen}
\usage{
read_screen(session, ...)
}
\arguments{
\item{session}{A rexpect_session.}

\item{...}{Other arguments passed on to \code{tmuxr::capture_pane}.}
}
\description{
Read screen
}
