% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexpect-package.R
\docType{package}
\name{rexpect-package}
\alias{rexpect}
\alias{rexpect-package}
\title{rexpect: Automate Interactions with Textual Interfaces}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

rexpect allows you to automate interactions with programs that
    expose a text terminal interface. The API is inspired by the original
    Expect tool by Don Libes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datascienceworkshops/rexpect}
  \item Report bugs at \url{https://github.com/datascienceworkshops/rexpect/issues}
}

}
\author{
\strong{Maintainer}: Jeroen Janssens \email{jeroen@datascienceworkshops.com} (\href{https://orcid.org/0000-0002-5878-8924}{ORCID})

Other contributors:
\itemize{
  \item Data Science Workshops [copyright holder, funder]
}

}
\keyword{internal}
