% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{send_keys}
\alias{send_keys}
\title{Send keys}
\usage{
send_keys(session, ..., literal = FALSE, count = 1L)
}
\arguments{
\item{session}{A rexpect_session.}

\item{...}{Strings. Keys to send.}

\item{literal}{A logical. If \code{TRUE}, key name lookup is disabled and the keys
are processed as literal UTF-8 characters. Default: \code{FALSE}.}

\item{count}{An integer. Number of times the keys are sent. Default: \code{1L}.}
}
\description{
Send keys
}
