% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{send_lines}
\alias{send_lines}
\title{Send literal text}
\usage{
send_lines(session, ..., wait = FALSE)
}
\arguments{
\item{session}{A rexpect_session.}

\item{...}{One or more strings.}

\item{wait}{A logical. If \code{TRUE}, wait for prompt to appear before sending
each line. Default: \code{FALSE}.}
}
\value{
session
}
\description{
Send literal text
}
