% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{send_script}
\alias{send_script}
\title{Send script}
\usage{
send_script(session, code, marker = "#!", ...)
}
\arguments{
\item{session}{A rexpect_session.}

\item{code}{A vector of strings.}

\item{marker}{A string. Default: \verb{#!}.}

\item{...}{Default values.}
}
\description{
Send script
}
