% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{set_prompt}
\alias{set_prompt}
\title{Set the prompt}
\usage{
set_prompt(session, prompt)
}
\arguments{
\item{session}{A rexpect_session.}

\item{prompt}{String containing a regular expression that matches all
relevant prompts.}
}
\description{
Set the prompt
}
