% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spawn.R
\name{spawn}
\alias{spawn}
\title{Spawn a new rexpect session}
\usage{
spawn(command, name = NULL, prompt = NULL, width = 80, height = 24)
}
\arguments{
\item{command}{A vector of strings.}

\item{name}{A string.}

\item{prompt}{A string.}

\item{width}{An integer.}

\item{height}{An integer.}
}
\value{
A rexpect_session.
}
\description{
Spawn a new rexpect session
}
