% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spawn.R
\name{wait}
\alias{wait}
\title{Wait}
\usage{
wait(session, duration = 0.1)
}
\arguments{
\item{session}{A rexpect_session.}

\item{duration}{A numerical. Time to wait in seconds. Default: \code{0.1}.}
}
\description{
Wait
}
