% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect.R
\name{expect_prompt}
\alias{expect_prompt}
\title{Expect prompt}
\usage{
expect_prompt(session, time = 0.1)
}
\arguments{
\item{session}{A rexpect_session.}

\item{time}{Numerical. Time to wait in seconds between tries.}
}
\description{
Block a session until the prompt appears.
}
